install Email::Address and Email::Address::List from cpan
RT 4.2.16 and 4.4.4 require newer version of these modules because of
security fixes. While the security fixes are included in the Debian
packages, the module versions where not changed.
So we need to install the modules from cpam to make the RT installer
happy.

We need to install Module::Install before Email::Address::List, because
this module requires Module::Install.